VA5

+++ app/Services/AuthService.php
@@ -3,22 +3,35 @@
 namespace App\Services;
 
 use App\Events\SignupEvent;
+use App\Http\Requests\Auth\loginAndRefreshRequest\loginRequest;
+use App\Http\Requests\Auth\loginAndRefreshRequest\refreshTokenRequest;
 use App\Http\Requests\Auth\signupAndVerificationRequests\checkVerificationCodeRequest;
 use App\Http\Requests\Auth\signupAndVerificationRequests\resendVerificationRequest;
 use App\Http\Requests\Auth\signupAndVerificationRequests\SignupRequest;
-use App\Listeners\SendEmailVerificationCodeListener;
 use App\Models\User;
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Ramsey\Uuid\Exception\InvalidArgumentException;
+use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
 
 class AuthService
 {
+    /**
+     * @param FileService $fileService
+     */
     public function __construct(public FileService $fileService)
     {
     }
 
-    public function signup(SignupRequest $request)
+    /**
+     * @param SignupRequest $request
+     * @return mixed
+     */
+    public function signup(SignupRequest $request): mixed
     {
         $validatedData = $request->validated(); // Get all validated data
 
@@ -41,23 +54,32 @@ class AuthService
 
     /**
      * @param checkVerificationCodeRequest $request
+     * @param $userVerificationCodeExpireTime
      * @return string
      * @throws Exception
      */
-    public function checkVerificationCode(checkVerificationCodeRequest $request , &$userVerificationCodeExpireTime): string
+    public function checkVerificationCode(checkVerificationCodeRequest $request, &$userVerificationCodeExpireTime): string
     {
-
-        if (!$user = User::where('email', $request->email)->first())
-            throw new  Exception('there is no user for this email');
+        $requestValidated = $request->validated();
+        if (!$user = User::where('email', $requestValidated['email'])->first())
+            throw new  ModelNotFoundException(
+                'there is no user for this email'
+            );
         if ($user->email_verified_at)
-            throw new  Exception('this email is used (verified) please inter another email ');
-        if ($user->verification_code != $request->verification_code)
-            throw new Exception('invalid user verification code ');
-        $userVerificationCodeExpireTime =Carbon::parse( $user->code_sent_at)->addMinutes(3);
+            throw new  UnprocessableEntityHttpException(
+                'this email is used (verified),
+                 please inter another email');
+        if ($user->verification_code != $requestValidated['verification_code'])
+            throw new UnprocessableEntityHttpException(
+                'invalid user verification code
+                ');
+        $userVerificationCodeExpireTime = Carbon::parse($user->code_sent_at)->addMinutes(3);
         if (Carbon::now()->isAfter($userVerificationCodeExpireTime)) {
-            throw new Exception('the code is expired , please  try to send the code again ');
+            throw new UnprocessableEntityHttpException(
+                'the code is expired , please  try to send the code again '
+            );
         } else {
-            $token = $user->createToken('my-api-token')->plainTextToken; // Replace with a descriptive token name
+            $token = $user->createToken('signup-token')->plainTextToken; // Replace with a descriptive token name
         }
         $user->email_verified_at = Carbon::now();
         $user->remember_token = $token;
@@ -66,14 +88,17 @@ class AuthService
     }
 
     /**
-     * @param $request
+     * @param refreshTokenRequest $request
      * @param $TokenExpireTime
      * @return string
      */
-    public function refreshUserToken($request, &$TokenExpireTime): string
+    public function refreshUserToken(refreshTokenRequest $request, &$TokenExpireTime): string
     {
+        $request = $request->validated();
         $TokenExpireTime = Carbon::now()->addMinutes(config('sanctum.refresh_expiration'));
-        $user = User::where('email', $request->email)->first();
+        if (!$user = User::where('email', $request->email)->first()) {
+            throw new  ModelNotFoundException('there is no user for this email');
+        }
         $user->tokens()->delete();
         $accessToken = $user->createToken(
             'refresh_token',
@@ -83,13 +108,50 @@ class AuthService
         return $accessToken->plainTextToken;
     }
 
+    /**
+     * @param resendVerificationRequest $request
+     * @return void
+     * @throws Exception
+     */
     public function resendVerificationCode(resendVerificationRequest $request): void
     {
         $userEmail = $request->email;
         if (!$user = User::where('email', $userEmail)->first())
-            throw new  Exception('there is no user for this email');
+            throw new  ModelNotFoundException('there is no user for this email');
         $user->verification_code = null;
         $user->code_sent_at = Carbon::now();
         event(new SignupEvent($user));
+    }
+
+    /**
+     * @param loginRequest $request
+     * @param $TokenExpireTime
+     * @return string
+     * @throws Exception|ModelNotFoundException |InvalidArgumentException
+     */
+    public function logInUser(loginRequest $request, &$TokenExpireTime): string
+    {
+        $validatedData = $request->validated(); // Access validated data from the request
+        if (!$user = User::where('email', $validatedData['email'])->first())
+            throw new ModelNotFoundException('there is no user for this email ');
+        if (!Auth::attempt($request->only('email', 'password')))
+            throw new InvalidArgumentException('Invalid credentials');
+        if (!$token = $user->createToken('login-token')->plainTextToken)
+            throw new Exception('there is some thing error please try again ');
+        $TokenExpireTime = Carbon::now()->addMinutes(config('sanctum.expiration') / 60);
+        $user->remember_token = $token;
+        $user->save();
+        return $token;
+    }
+
+    /**
+     * @param $request
+     * @return void
+     */
+    public function logoutUser($request)
+    {
+
+        auth()->user()->tokens()->delete(); // Revoke all user tokens
+        return;
     }
 }
